Revert "Jebanyrefactor"
