+++ data_manager.py
@@ -1,5 +1,57 @@
+#rozne funkcje:
+#wpisywanie:
+#czasu
+#id
+#liczenie vote'ow
+#wyswietlenia
+#obrazek
+#wszystko musi skladac i rozkladac z list/ w liste
+
 import connection
-import data_manager_questions
+import uuid
+from datetime import datetime
+import os
+
+
+def get_all_answers():
+    a_list = connection.import_data('ask-mate-python/sample_data/answer.csv')
+    return a_list
+
+
+def get_questions():
+    q_list = connection.import_data('ask-mate-python/sample_data/question.csv')
+    return q_list
+
+
+def add_answer(form_data, id):
+    answers = connection.import_data('ask-mate-python/sample_data/answer.csv')
+    new_answer = {
+        'id': uuid.uuid4(),
+        'submission_time': datetime.now(),
+        'vote_number': 0,
+        'question_id': id,
+        'message': form_data['answer'],
+        'image': None
+    }
+    answers.append(new_answer)
+    connection.write_file(answers, 'ask-mate-python/sample_data/answer.csv')
+
+
+def one_question(id):
+    questions = get_questions()
+    for question in questions:
+        if question['id'] == id:
+            return question
+
+
+def get_answers_to_question(id):
+    id = id
+    answers = get_all_answers()
+    filtered_answers = []
+    for answer in answers:
+        if answer['question_id'] == id:
+            filtered_answers.append(answer)
+    return filtered_answers
 
 
 def delete_element(element_type, element_id):
@@ -15,6 +67,7 @@ def delete_element(element_type, element_id):
     updated_data = [data_element for data_element in data if data_element['id'] != element_id]
     connection.write_file(updated_data, f'ask-mate-python/sample_data/{element_type}.csv')
 
+    answers = []
     if element_type == "question":
         answers = connection.import_data('ask-mate-python/sample_data/answer.csv')
 
@@ -27,7 +80,7 @@ def delete_element(element_type, element_id):
 
 
 def sorting_questions(filter, descending):
-    questions = data_manager_questions.get_questions()
+    questions = get_questions()
     sorting_key = 'vote_number'
     if filter == "by_date":
         sorting_key = 'submission_time'
@@ -35,7 +88,7 @@ def sorting_questions(filter, descending):
 
 
 def checker(type):
-    questions = data_manager_questions.get_questions()
+    questions = get_questions()
     if type == 'submission_time':
         return True if questions[0][type] < questions[-1][type] else False
     else:
@@ -56,3 +109,19 @@ def update_image(file_type, filename, id):
         if element['id'] == id:
             element['image'] = '/static/images/' + filename
     connection.write_file(data, f"ask-mate-python/sample_data/{file_type}.csv")
+
+
+def update_question(id, title, description, file_type):
+    questions = connection.import_data('ask-mate-python/sample_data/question.csv')
+    for question in questions:
+        if question['id'] == id:
+            question['title'] = title
+            question['message'] = description
+    connection.write_file(questions, f"ask-mate-python/sample_data/{file_type}.csv")
+
+
+def newest_question():
+    new_question = []
+    sorted_questions = sorting_questions("by_date", True)
+    new_question.append(sorted_questions[0])
+    return new_question
